сделал адаптер для почасовой температуры и добавил иконки

+++ app/src/main/java/com/monsieur/cloy/weatherapp/api/classes/Minutely.kt
@@ -9,7 +9,7 @@ class Minutely(
     var dt: Int,//Время прогнозируемых данных, unix, UTC
     @SerializedName("precipitation")
     @Expose
-    var precipitation: Int//Объем осадков, мм
+    var precipitation: Double//Объем осадков, мм
 ) {
 
 }
